requirements.txt
torch
torchvision
safetensors
git+https://github.com/huggingface/diffusers.git
transformers
lycoris-lora==1.8.3
flatten_json
pyyaml
oyaml
tensorboard
kornia
invisible-watermark
einops
accelerate
toml
albumentations
pydantic
omegaconf
k-diffusion
open_clip_torch
timm
prodigyopt
controlnet_aux==0.0.7
python-dotenv
bitsandbytes
hf_transfer
lpips
pytorch_fid
optimum-quanto
sentencepiece
huggingface_hub
peft
gradio
python-slugify
python-decouple
boto3
# mediapipe
google-cloud-storage